data_generator/requirements.txt
pandas
numpy
Faker
PyYAML
tqdm
python-dateutil
